:(go.mod): included new modules
