+++ go.mod
@@ -1,0 +1,5 @@
+module GO-GET-SOME-STYLEZ
+
+go 1.24.4
+
+require github.com/common-nighthawk/go-figure v0.0.0-20210622060536-734e95fb86be // indirect
